+++ tests/monitoring/monitoring.go
@@ -23,8 +23,6 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"regexp"
-	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -32,17 +30,11 @@ import (
 
 	promv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 
-	rbacv1 "k8s.io/api/rbac/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/apimachinery/pkg/util/rand"
-
 	v1 "kubevirt.io/api/core/v1"
 	"kubevirt.io/client-go/kubecli"
 
 	"kubevirt.io/kubevirt/tests"
-	"kubevirt.io/kubevirt/tests/clientcmd"
 	"kubevirt.io/kubevirt/tests/decorators"
 	"kubevirt.io/kubevirt/tests/flags"
 	"kubevirt.io/kubevirt/tests/framework/checks"
@@ -52,169 +44,10 @@ import (
 	"kubevirt.io/kubevirt/tests/util"
 )
 
-type alerts struct {
-	deploymentName       string
-	downAlert            string
-	noReadyAlert         string
-	restErrorsBurtsAlert string
-	restErrorsHighAlert  string
-	lowCountAlert        string
-}
-
-var (
-	virtApi = alerts{
-		deploymentName:       "virt-api",
-		downAlert:            "VirtAPIDown",
-		restErrorsBurtsAlert: "VirtApiRESTErrorsBurst",
-		restErrorsHighAlert:  "VirtApiRESTErrorsHigh",
-		lowCountAlert:        "LowVirtAPICount",
-	}
-	virtController = alerts{
-		deploymentName:       "virt-controller",
-		downAlert:            "VirtControllerDown",
-		noReadyAlert:         "NoReadyVirtController",
-		restErrorsBurtsAlert: "VirtControllerRESTErrorsBurst",
-		restErrorsHighAlert:  "VirtControllerRESTErrorsHigh",
-		lowCountAlert:        "LowVirtControllersCount",
-	}
-	virtHandler = alerts{
-		deploymentName:       "virt-handler",
-		restErrorsBurtsAlert: "VirtHandlerRESTErrorsBurst",
-		restErrorsHighAlert:  "VirtHandlerRESTErrorsHigh",
-	}
-	virtOperator = alerts{
-		deploymentName:       "virt-operator",
-		downAlert:            "VirtOperatorDown",
-		noReadyAlert:         "NoReadyVirtOperator",
-		restErrorsBurtsAlert: "VirtOperatorRESTErrorsBurst",
-		restErrorsHighAlert:  "VirtOperatorRESTErrorsHigh",
-		lowCountAlert:        "LowVirtOperatorCount",
-	}
-)
-
 var _ = Describe("[Serial][sig-monitoring]Monitoring", Serial, decorators.SigMonitoring, func() {
-
 	var err error
 	var virtClient kubecli.KubevirtClient
-	var scales *Scaling
 	var prometheusRule *promv1.PrometheusRule
-
-	checkAlert := func(alertName string) error {
-		alerts, err := getAlerts(virtClient)
-		if err != nil {
-			return err
-		}
-		for _, alert := range alerts {
-			if string(alert.Labels["alertname"]) == alertName {
-				return nil
-			}
-		}
-		return fmt.Errorf("alert doesn't exist: %v", alertName)
-	}
-
-	verifyAlertExistWithCustomTime := func(alertName string, timeout time.Duration) {
-		Eventually(func() error {
-			return checkAlert(alertName)
-		}, timeout, 10*time.Second).Should(BeNil())
-	}
-
-	verifyAlertExist := func(alertName string) {
-		verifyAlertExistWithCustomTime(alertName, 120*time.Second)
-	}
-
-	waitUntilAlertDoesNotExist := func(alertName string) {
-		Eventually(func() error {
-			alerts, err := getAlerts(virtClient)
-			if err != nil {
-				return err
-			}
-			for _, alert := range alerts {
-				if alertName == string(alert.Labels["alertname"]) {
-					return fmt.Errorf("alert exist: %v", alertName)
-				}
-			}
-			return nil
-		}, 5*time.Minute, 1*time.Second).Should(BeNil())
-	}
-
-	increaseRateLimit := func() {
-		rateLimitConfig := &v1.ReloadableComponentConfiguration{
-			RestClient: &v1.RESTClientConfiguration{
-				RateLimiter: &v1.RateLimiter{
-					TokenBucketRateLimiter: &v1.TokenBucketRateLimiter{
-						Burst: 300,
-						QPS:   300,
-					},
-				},
-			},
-		}
-		originalKubeVirt := util.GetCurrentKv(virtClient)
-		originalKubeVirt.Spec.Configuration.ControllerConfiguration = rateLimitConfig
-		originalKubeVirt.Spec.Configuration.HandlerConfiguration = rateLimitConfig
-		tests.UpdateKubeVirtConfigValueAndWait(originalKubeVirt.Spec.Configuration)
-	}
-
-	updatePromRules := func(newRules *promv1.PrometheusRule) {
-		err = virtClient.
-			PrometheusClient().MonitoringV1().
-			PrometheusRules(flags.KubeVirtInstallNamespace).
-			Delete(context.Background(), "prometheus-kubevirt-rules", metav1.DeleteOptions{})
-
-		Expect(err).ToNot(HaveOccurred())
-
-		_, err = virtClient.
-			PrometheusClient().MonitoringV1().
-			PrometheusRules(flags.KubeVirtInstallNamespace).
-			Create(context.Background(), newRules, metav1.CreateOptions{})
-
-		Expect(err).ToNot(HaveOccurred())
-	}
-
-	getPrometheusAlerts := func() promv1.PrometheusRule {
-		promRules, err := virtClient.
-			PrometheusClient().MonitoringV1().
-			PrometheusRules(flags.KubeVirtInstallNamespace).
-			Get(context.Background(), "prometheus-kubevirt-rules", metav1.GetOptions{})
-		Expect(err).ToNot(HaveOccurred())
-
-		var newRules promv1.PrometheusRule
-		promRules.DeepCopyInto(&newRules)
-
-		newRules.Annotations = nil
-		newRules.ObjectMeta.ResourceVersion = ""
-		newRules.ObjectMeta.UID = ""
-
-		return newRules
-	}
-
-	reduceAlertPendingTime := func() {
-		By("Reducing alert pending time")
-		newRules := getPrometheusAlerts()
-		var re = regexp.MustCompile("\\[\\d+m\\]")
-
-		var gs []promv1.RuleGroup
-		for _, group := range newRules.Spec.Groups {
-			var rs []promv1.Rule
-			for _, rule := range group.Rules {
-				var r promv1.Rule
-				rule.DeepCopyInto(&r)
-				if r.Alert != "" {
-					r.For = "0m"
-					r.Expr = intstr.FromString(re.ReplaceAllString(r.Expr.String(), `[1m]`))
-					r.Expr = intstr.FromString(strings.ReplaceAll(r.Expr.String(), ">= 300", ">= 0"))
-				}
-				rs = append(rs, r)
-			}
-
-			gs = append(gs, promv1.RuleGroup{
-				Name:  group.Name,
-				Rules: rs,
-			})
-		}
-		newRules.Spec.Groups = gs
-
-		updatePromRules(&newRules)
-	}
 
 	BeforeEach(func() {
 		virtClient = kubevirt.Client()
@@ -249,148 +82,6 @@ var _ = Describe("[Serial][sig-monitoring]Monitoring", Serial, decorators.SigMon
 		})
 	})
 
-	Context("Up metrics", func() {
-		BeforeEach(func() {
-			scales = NewScaling(virtClient, []string{virtOperator.deploymentName, virtController.deploymentName, virtApi.deploymentName})
-			scales.UpdateScale(virtOperator.deploymentName, int32(0))
-
-			reduceAlertPendingTime()
-		})
-
-		AfterEach(func() {
-			scales.RestoreAllScales()
-
-			time.Sleep(10 * time.Second)
-			alerts := []string{
-				virtOperator.downAlert, virtOperator.noReadyAlert, virtOperator.lowCountAlert,
-				virtController.downAlert, virtController.noReadyAlert, virtController.lowCountAlert,
-				virtApi.downAlert, virtApi.noReadyAlert, virtApi.lowCountAlert,
-			}
-			for _, alert := range alerts {
-				waitUntilAlertDoesNotExist(alert)
-			}
-		})
-
-		It("VirtOperatorDown and NoReadyVirtOperator should be triggered when virt-operator is down", func() {
-			verifyAlertExist(virtOperator.downAlert)
-			verifyAlertExist(virtOperator.noReadyAlert)
-		})
-
-		It("LowVirtOperatorCount should be triggered when virt-operator count is low", decorators.RequiresTwoSchedulableNodes, func() {
-			verifyAlertExist(virtOperator.lowCountAlert)
-		})
-
-		It("VirtControllerDown and NoReadyVirtController should be triggered when virt-controller is down", func() {
-			scales.UpdateScale(virtController.deploymentName, int32(0))
-			verifyAlertExist(virtController.downAlert)
-			verifyAlertExist(virtController.noReadyAlert)
-		})
-
-		It("LowVirtControllersCount should be triggered when virt-controller count is low", decorators.RequiresTwoSchedulableNodes, func() {
-			scales.UpdateScale(virtController.deploymentName, int32(0))
-			verifyAlertExist(virtController.lowCountAlert)
-		})
-
-		It("VirtApiDown should be triggered when virt-api is down", func() {
-			scales.UpdateScale(virtApi.deploymentName, int32(0))
-			verifyAlertExist(virtApi.downAlert)
-		})
-
-		It("LowVirtAPICount should be triggered when virt-api count is low", decorators.RequiresTwoSchedulableNodes, func() {
-			scales.UpdateScale(virtApi.deploymentName, int32(0))
-			verifyAlertExist(virtApi.lowCountAlert)
-		})
-	})
-
-	Context("Errors metrics", func() {
-		var crb *rbacv1.ClusterRoleBinding
-
-		BeforeEach(func() {
-			virtClient = kubevirt.Client()
-
-			crb, err = virtClient.RbacV1().ClusterRoleBindings().Get(context.Background(), "kubevirt-operator", metav1.GetOptions{})
-			util.PanicOnError(err)
-
-			increaseRateLimit()
-
-			scales = NewScaling(virtClient, []string{virtOperator.deploymentName})
-			scales.UpdateScale(virtOperator.deploymentName, int32(0))
-
-			reduceAlertPendingTime()
-		})
-
-		AfterEach(func() {
-			crb.Annotations = nil
-			crb.ObjectMeta.ResourceVersion = ""
-			crb.ObjectMeta.UID = ""
-			_, err = virtClient.RbacV1().ClusterRoleBindings().Create(context.Background(), crb, metav1.CreateOptions{})
-			if !errors.IsAlreadyExists(err) {
-				util.PanicOnError(err)
-			}
-			scales.RestoreAllScales()
-
-			time.Sleep(10 * time.Second)
-			waitUntilAlertDoesNotExist(virtOperator.downAlert)
-			waitUntilAlertDoesNotExist(virtApi.downAlert)
-			waitUntilAlertDoesNotExist(virtController.downAlert)
-			waitUntilAlertDoesNotExist(virtHandler.downAlert)
-		})
-
-		It("VirtApiRESTErrorsBurst and VirtApiRESTErrorsHigh should be triggered when requests to virt-api are failing", func() {
-			randVmName := rand.String(6)
-
-			Eventually(func(g Gomega) {
-				cmd := clientcmd.NewVirtctlCommand("vnc", randVmName)
-				err := cmd.Execute()
-				Expect(err).To(HaveOccurred())
-
-				g.Expect(checkAlert(virtApi.restErrorsBurtsAlert)).To(Not(HaveOccurred()))
-				g.Expect(checkAlert(virtApi.restErrorsHighAlert)).To(Not(HaveOccurred()))
-			}, 5*time.Minute, 500*time.Millisecond).Should(Succeed())
-		})
-
-		It("VirtOperatorRESTErrorsBurst and VirtOperatorRESTErrorsHigh should be triggered when requests to virt-operator are failing", func() {
-			scales.RestoreScale(virtOperator.deploymentName)
-			err = virtClient.RbacV1().ClusterRoleBindings().Delete(context.Background(), crb.Name, metav1.DeleteOptions{})
-			Expect(err).ToNot(HaveOccurred())
-
-			Eventually(func(g Gomega) {
-				g.Expect(checkAlert(virtOperator.restErrorsBurtsAlert)).To(Not(HaveOccurred()))
-				g.Expect(checkAlert(virtOperator.restErrorsHighAlert)).To(Not(HaveOccurred()))
-			}, 5*time.Minute, 500*time.Millisecond).Should(Succeed())
-		})
-
-		It("VirtControllerRESTErrorsBurst and VirtControllerRESTErrorsHigh should be triggered when requests to virt-controller are failing", func() {
-			err = virtClient.RbacV1().ClusterRoleBindings().Delete(context.Background(), "kubevirt-controller", metav1.DeleteOptions{})
-			Expect(err).ToNot(HaveOccurred())
-
-			vmi := tests.NewRandomVMI()
-
-			Eventually(func(g Gomega) {
-				_, _ = virtClient.VirtualMachineInstance(util.NamespaceTestDefault).Create(context.Background(), vmi)
-				_ = virtClient.VirtualMachineInstance(util.NamespaceTestDefault).Delete(context.Background(), vmi.Name, &metav1.DeleteOptions{})
-
-				g.Expect(checkAlert(virtController.restErrorsBurtsAlert)).To(Not(HaveOccurred()))
-				g.Expect(checkAlert(virtController.restErrorsHighAlert)).To(Not(HaveOccurred()))
-			}, 5*time.Minute, 500*time.Millisecond).Should(Succeed())
-		})
-
-		It("VirtHandlerRESTErrorsBurst and VirtHandlerRESTErrorsHigh should be triggered when requests to virt-handler are failing", func() {
-			err = virtClient.RbacV1().ClusterRoleBindings().Delete(context.Background(), "kubevirt-handler", metav1.DeleteOptions{})
-			Expect(err).ToNot(HaveOccurred())
-
-			vmi := tests.NewRandomVMI()
-
-			Eventually(func(g Gomega) {
-				_, _ = virtClient.VirtualMachineInstance(util.NamespaceTestDefault).Create(context.Background(), vmi)
-				_ = virtClient.VirtualMachineInstance(util.NamespaceTestDefault).Delete(context.Background(), vmi.Name, &metav1.DeleteOptions{})
-
-				g.Expect(checkAlert(virtHandler.restErrorsBurtsAlert)).To(Not(HaveOccurred()))
-				g.Expect(checkAlert(virtHandler.restErrorsHighAlert)).To(Not(HaveOccurred()))
-			}, 5*time.Minute, 500*time.Millisecond).Should(Succeed())
-		})
-	})
-
 	Context("Migration Alerts", decorators.SigComputeMigrations, func() {
 		It("KubeVirtVMIExcessiveMigrations should be triggered when a VMI has been migrated more than 12 times during the last 24 hours", func() {
 			By("Starting the VirtualMachineInstance")
@@ -408,7 +99,7 @@ var _ = Describe("[Serial][sig-monitoring]Monitoring", Serial, decorators.SigMon
 			}
 
 			By("Verifying KubeVirtVMIExcessiveMigration alert exists")
-			verifyAlertExist("KubeVirtVMIExcessiveMigrations")
+			verifyAlertExist(virtClient, "KubeVirtVMIExcessiveMigrations")
 
 			// delete VMI
 			By("Deleting the VMI")
@@ -473,11 +164,11 @@ var _ = Describe("[Serial][sig-monitoring]Monitoring", Serial, decorators.SigMon
 			kv := disableVirtHandler()
 
 			By("Verifying KubeVirtNoAvailableNodesToRunVMs alert exists if emulation is disabled")
-			verifyAlertExistWithCustomTime("KubeVirtNoAvailableNodesToRunVMs", 10*time.Minute)
+			verifyAlertExistWithCustomTime(virtClient, "KubeVirtNoAvailableNodesToRunVMs", 10*time.Minute)
 
 			By("Restoring virt-handler")
 			restoreVirtHandler(kv)
-			waitUntilAlertDoesNotExist("KubeVirtNoAvailableNodesToRunVMs")
+			waitUntilAlertDoesNotExist(virtClient, "KubeVirtNoAvailableNodesToRunVMs")
 		})
 	})
